+++ diesel-wasm-sqlite/src/query_builder/returning.rs
@@ -1,4 +1,4 @@
-use crate::storage::wasm_sqlite::backend::{SqliteReturningClause, WasmSqlite};
+use crate::backend::{SqliteReturningClause, WasmSqlite};
 // use diesel::backend::Backend;
 use diesel::query_builder::ReturningClause;
 use diesel::query_builder::{AstPass, QueryFragment};
